requirements.txt
faster-whisper
google-generativeai
requests
opencv-python
pillow
subprocess
pyshine
pyyaml